chore: refactor context
