+++ playground/src/hooks/usePanda.ts
@@ -1,94 +1,13 @@
+import { usePandaContext } from '@/src/hooks/usePandaContext'
 import { State } from '@/src/hooks/usePlayground'
-import { getResolvedConfig } from '@/src/lib/resolve-config'
-import { Generator } from '@pandacss/generator'
 import { createProject } from '@pandacss/parser'
-import presetBase from '@pandacss/preset-base'
-import presetTheme from '@pandacss/preset-panda'
-import { Config, Preset, StaticCssOptions } from '@pandacss/types'
-import { createHooks } from 'hookable'
-import { merge } from 'merge-anything'
-import { useMemo, useRef } from 'react'
+import { Config } from '@pandacss/types'
+import { useMemo } from 'react'
 
-const playgroundPreset: Preset = {
-  theme: {
-    recipes: {
-      playgroundError: {
-        className: 'playgroundError',
-        base: {
-          p: '2',
-          color: 'red.400',
-          display: 'flex',
-          gap: '2',
-          alignItems: 'center',
-          background: { base: 'white', _dark: '#262626' },
-          borderBottomWidth: '1px',
-          borderBottomColor: { base: 'gray.100', _dark: '#262626' },
-          textStyle: 'sm',
-
-          '& > span': {
-            borderRadius: 'full',
-            display: 'flex',
-            padding: '1',
-            alignItems: 'center',
-            background: {
-              base: 'rgba(235, 94, 66, 0.2)',
-              _dark: 'rgba(235, 94, 66, 0.1)',
-            },
-          },
-
-          '& svg': {
-            h: '16px',
-            w: '16px',
-          },
-        },
-        variants: {
-          style: {
-            empty: {},
-          },
-        },
-      },
-    },
-  },
-}
-
-export function usePanda(state: State, userConfig: Config | null) {
+export function usePanda(state: State, config: Config | null) {
   const { code: source, css } = state
-  const previousContext = useRef<Generator | null>(null)
 
-  const context = useMemo(() => {
-    const { presets, ...restConfig } = userConfig ?? {}
-
-    const config = getResolvedConfig({
-      cwd: '',
-      include: [],
-      outdir: 'styled-system',
-      preflight: true,
-      optimize: true,
-      presets: [presetBase, presetTheme, playgroundPreset, ...(presets ?? [])],
-      ...restConfig,
-      staticCss: merge(restConfig.staticCss, {
-        recipes: { playgroundError: ['*'] } as StaticCssOptions['recipes'],
-      }),
-
-      jsxFramework: restConfig.jsxFramework ? 'react' : undefined,
-    })
-
-    try {
-      // in event of error (invalid token format), use previous generator
-      const context = new Generator({
-        dependencies: [],
-        serialized: '',
-        deserialize: () => config!,
-        path: '',
-        hooks: createHooks(),
-        config: config as any,
-      })
-      previousContext.current = context
-      return context
-    } catch {
-      return previousContext.current!
-    }
-  }, [userConfig])
+  const context = usePandaContext(config)
 
   const staticArtifacts = useMemo(() => {
     context.appendLayerParams()
